Refactor : forecastchart 분리
