+++ src/routes/MainPage/HealthCharts/CostForecastChart/index.tsx
@@ -1,14 +1,9 @@
 import { VictoryAxis, VictoryBar, VictoryChart, VictoryLine, VictoryScatter } from 'victory'
 import { CallbackArgs } from 'victory-core'
 
-import styles from './forecastChart.module.scss'
+import styles from './costForecastChart.module.scss'
 
-const ForecastChart = () => {
-  const compareScoreDecade = [
-    { x: '나', HealthScore: 875 },
-    { x: '10년 후', HealthScore: 790 },
-  ]
-
+const CostForecastChart = () => {
   const compareMediExpenseDecade = [
     { x: '나', mediExpense: 93335 },
     { x: '10년 후', mediExpense: 129455 },
@@ -19,40 +14,8 @@ const ForecastChart = () => {
       <VictoryChart height={300}>
         <VictoryAxis style={{ axis: { stroke: 'white' }, tickLabels: { fontSize: 20 } }} />
         <VictoryBar
-          barWidth={60}
-          data={compareScoreDecade}
-          x='x'
-          y='HealthScore'
-          labels={({ datum }) => `${datum.HealthScore}점`}
-          style={{
-            data: {
-              fill: ({ datum }) => (datum.x === '나' ? '#ffd300' : '#fe612c'),
-            },
-            labels: {
-              fill: ({ datum }: CallbackArgs) => (datum.x === '나' ? '#fe612c' : 'black'),
-              fontSize: 20,
-            },
-          }}
-        />
-        <VictoryLine data={compareScoreDecade} x='x' y='HealthScore' />
-        <VictoryScatter
-          data={compareScoreDecade}
-          x='x'
-          y='HealthScore'
-          style={{
-            data: {
-              fill: ({ datum }) => (datum.x === '나' ? 'grey' : 'white'),
-              stroke: ({ datum }) => (datum.x === '나' ? 'none' : 'black'),
-              strokeWidth: 1,
-            },
-          }}
-          size={5}
-        />
-      </VictoryChart>
-      <VictoryChart domainPadding={60}>
-        <VictoryAxis style={{ axis: { stroke: 'white' }, tickLabels: { fontSize: 20 } }} />
-        <VictoryBar
           data={compareMediExpenseDecade}
+          barWidth={60}
           x='x'
           y='mediExpense'
           labels={({ datum }) => datum.mediExpense}
@@ -85,4 +48,4 @@ const ForecastChart = () => {
   )
 }
 
-export default ForecastChart
+export default CostForecastChart
