+++ src/routes/MainPage/HealthCharts/HealthForecastChart/healthForecastChart.module.scss
@@ -1,0 +1,5 @@
+.container {
+  width: 330px;
+  height: 150px;
+  margin-top: 50px;
+}
